refactor: reorg rest of package
