+++ cognite/powerops/_version.py
@@ -1,1 +1,1 @@
-__version__ = "0.24.0"
+__version__ = "0.25.0"
