+++ cognite/powerops/data_models/CogShop.gql
